rename Npart and start splitting functions
